Add a header component. Change size of arrow
Add the header component and place all the stuff for this here (earlier in main-frame). Change the size of the arrow so it's correct and responsive, but only responsive to the parent - the height of the parent-div is now hard coded. Change the clock to smaller text and right align.