Differentiate between current step and other steps
This commit will make it so that there is always one currently active HLR step. When the user presses one of the analysis buttons, only the remaining ones are now changed according to this, while the previous steps remain unchanged. After this, the flow moves on to the next step by incrementing the currentStepIndex variable.

Not yet implemented: the user can still change the internal state of the non-current steps. For example, when pressing the analysis buttons on another step than the current, the button will change color even though it does not affect the application itself. This should be prevented in the future as it is rather non-intuitive.