Change the assignIndex variable to a static variable in the Step class
Each Step object needs to be assigned an incremented ID (an index). Before this
commit, this value was manually incremented as it was assigned to each step in t
heir constructor. This is now changed to a static variable in the Step class - doing this makes it impossible to accidentally forget to increment the variable, preventing steps with the same index from happening.