Implements type-changing in barnhlrsettings, also overhauls the keypad.
This commit modifies the entire keypad functionality, making it emit strings instead of number and complete removes its sanity checking. This is now handled by the barnhlrsettings instead, which makes it possible to update after any change has been made to the keypad. All this has led to almost a complete rewrite of the barnhlrsettings functionality and keypad functionality.