Make ASSIGN_INDEX private
ASSIGN_INDEX should be private for safety reasons (nobody should be able to change it as it would lead to inconsistencies in the flow) but it still needs to be reset after a flow is finished. This commit introduces a reset method for that purpose, allowing the variable to be private.